add initial sensor code
