+++ custom_components/gpio_integration/light.py
@@ -15,10 +15,10 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .config_schema import PwmConfig
 from .core import DOMAIN, GpioEffect, StoppableThread, get_logger
 from .gpio.pin_factory import create_pin
 from .hub import Hub
+from .schemas.pwm import PwmConfig
 
 _LOGGER = get_logger()
 HIGH_BRIGHTNESS = 255
