+++ custom_components/gpio_integration/manifest.json
@@ -7,5 +7,5 @@
   "iot_class": "local_polling",
   "issue_tracker": "https://github.com/rosenkolev/home-assistant-gpio-integration/issues",
   "requirements": ["pigpio==1.78"],
-  "version": "1.1.0"
+  "version": "1.2.0"
 }
