+++ tests/mocked_modules.py
@@ -16,6 +16,7 @@ sys.modules["homeassistant.exceptions"] = Mock()
 sys.modules["homeassistant.components"] = Mock()
 sys.modules["homeassistant.components.cover"] = Mock()
 sys.modules["homeassistant.components.binary_sensor"] = Mock()
+sys.modules["homeassistant.components.sensor"] = Mock()
 sys.modules["homeassistant.components.switch"] = Mock()
 sys.modules["pigpio"] = Mock()
 sys.modules["RPi"] = Mock()
